app-software-dev-portfolio/Projects/document-cleaner-app/requirements.txt
streamlit
python-docx
PyPDF2
pandas 

🚀 Major Enhancements Implemented
1. Multi-Format Support
✅ PDF files - Extract text from PDF documents
✅ DOCX files - Process Word documents
✅ TXT/MD files - Handle text and markdown files
✅ Smart file detection - Automatically detects file type

2. Advanced Cleaning Options
✅ Configurable cleaning rules - Toggle different cleaning options
✅ URL removal - Remove web links
✅ Email removal - Strip email addresses
✅ Whitespace normalization - Clean up extra spaces
✅ YAML frontmatter removal - Optional YAML stripping

3. Enhanced User Interface
✅ Progress bars - Visual feedback during processing
✅ Statistics dashboard - Show file size, reduction %, chunk count
✅ Sidebar configuration - Easy access to all options
✅ File preview - Preview cleaned content before download
✅ Responsive layout - Better use of screen space

4. Advanced Features
✅ Multiple output formats - MD, TXT, CSV options
✅ Multiple output formats - MD, TXT, CSV options
✅ Flexible splitting - By bytes or character count
✅ CSV export - Export chunks as structured data
✅ Custom chunk sizes - Adjustable splitting parameters
✅ Better file naming - Include original filename in output

5. Professional Features
✅ Error handling - Graceful PDF processing errors
✅ File size display - Show upload file information
✅ Processing status - Real-time status updates
✅ Multiple download options - ZIP and CSV formats

New Capabilities
PDF Processing: Extract text from PDF documents
Configurable Cleaning: Choose what to remove (URLs, emails, etc.)
Flexible Splitting: Split by file size or character count
Multiple Outputs: Export as markdown, text, or CSV
Progress Tracking: Visual progress bars and status updates
Statistics: See processing results and file reduction
Preview: View cleaned content before downloading
